refactor(server): isolate websocket sm registration aftermath
Move the XEP-0198 post-registration resume-claim and fresh-bind invalidation logic out of the websocket transport loop and into the stream-management module. The connection loop still owns frame transport and registry publication, while stream_management now owns the claim completion outcomes and exact SM failure/resume wire responses.

Constraint: #350 asks for focused websocket boundary extraction without changing endpoint or XMPP behavior

Rejected: Split the whole websocket route in one PR | too broad for safe review after the #459 merge chain

Confidence: high

Scope-risk: narrow

Directive: Keep later #350 slices behavior-preserving and module-boundary focused

Tested: cargo fmt

Tested: cargo test -p waddle-server server::routes::websocket

Tested: cargo test -p waddle-xmpp --test xep0198_session_registry

Tested: cargo clippy -p waddle-server --tests -- -D warnings

Tested: git diff --check
